#134 - Extract AddWriter
