+++ src/main/java/com/artipie/helm/ChartsWriter.java
@@ -27,7 +27,6 @@ import com.artipie.asto.Key;
 import com.artipie.asto.Storage;
 import com.artipie.helm.metadata.Index;
 import com.artipie.helm.metadata.IndexYaml;
-import com.artipie.helm.metadata.IndexYamlMapping;
 import com.artipie.helm.metadata.ParsedChartName;
 import com.artipie.helm.metadata.YamlWriter;
 import com.artipie.helm.misc.DateTimeNow;
@@ -47,7 +46,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
-import org.apache.commons.lang3.tuple.Pair;
 
 /**
  * Writer of index file which combines source existed index file and add
@@ -92,80 +90,6 @@ final class ChartsWriter {
      */
     ChartsWriter(final Storage storage) {
         this.storage = storage;
-    }
-
-    /**
-     * Add info about charts to index. If index contains a chart with the same
-     * version, the exception will be generated. It has the next implementation.
-     * Read index file line by line. If we are in the `entries:` section, we will check
-     * whether the line is a name of chart (e.g. line has correct indent and ends
-     * with colon). It copy source index file line by line and if the line with
-     * version is met, the existence of this version in packages would be checked
-     * to avoid adding existed package. If the new name of chart is met, it will
-     * write remained versions from packages. When we read next line after end of
-     * `entries:` section from source index, we write info about remained charts
-     * in packages.
-     * @param source Path to temporary file with index
-     * @param out Path to temporary file in which new index would be written
-     * @param pckgs Packages collection which contains info about passed packages for
-     *  adding to index file. There is a version and chart yaml for each package.
-     * @return Result of completion
-     */
-    public CompletionStage<Void> addChartsToIndex(
-        final Path source,
-        final Path out,
-        final Map<String, Set<Pair<String, ChartYaml>>> pckgs
-    ) {
-        return this.storage.exists(IndexYaml.INDEX_YAML)
-            .thenCompose(this::versionsByPckgs)
-            .thenCompose(
-                vrsns -> {
-                    try (
-                        BufferedReader br = new BufferedReader(
-                            new InputStreamReader(Files.newInputStream(source))
-                        );
-                        BufferedWriter bufw = new BufferedWriter(
-                            new OutputStreamWriter(Files.newOutputStream(out))
-                        )
-                    ) {
-                        String line;
-                        boolean entrs = false;
-                        String name = null;
-                        YamlWriter writer = new YamlWriter(bufw, 2);
-                        while ((line = br.readLine()) != null) {
-                            final String trimmed = line.trim();
-                            final int lastposspace = ChartsWriter.lastPosOfSpaceInBegin(line);
-                            if (!entrs) {
-                                entrs = trimmed.equals(ChartsWriter.ENTRS);
-                            }
-                            if (entrs && new ParsedChartName(line).valid()) {
-                                if (name == null) {
-                                    writer = new YamlWriter(bufw, lastposspace);
-                                }
-                                if (lastposspace == writer.indent()) {
-                                    ChartsWriter.writeRemainedVersionsOfChart(name, pckgs, writer);
-                                    name = trimmed.replace(":", "");
-                                }
-                            }
-                            if (entrs) {
-                                ChartsWriter.throwIfVersionExists(trimmed, name, pckgs);
-                            }
-                            if (entrs && name != null && lastposspace == 0) {
-                                ChartsWriter.writeRemainedVersionsOfChart(name, pckgs, writer);
-                                ChartsWriter.writeRemainedChartsAfterCopyIndex(pckgs, writer);
-                                entrs = false;
-                            }
-                            ChartsWriter.writeAndReplaceTagGenerated(line, writer);
-                        }
-                        if (entrs) {
-                            ChartsWriter.writeRemainedChartsAfterCopyIndex(pckgs, writer);
-                        }
-                    } catch (final IOException exc) {
-                        throw new UncheckedIOException(exc);
-                    }
-                    return CompletableFuture.allOf();
-                }
-            );
     }
 
     /**
@@ -355,89 +279,6 @@ final class ChartsWriter {
             res = CompletableFuture.completedFuture(new HashMap<>());
         }
         return res;
-    }
-
-    /**
-     * Generates an exception if version of chart which contains in trimmed
-     * line exists in packages.
-     * @param trimmed Trimmed line from index file
-     * @param name Name of chart
-     * @param pckgs Packages collection which contains info about passed packages for
-     *  adding to index file. There is a version and chart yaml for each package.
-     */
-    private static void throwIfVersionExists(
-        final String trimmed,
-        final String name,
-        final Map<String, Set<Pair<String, ChartYaml>>> pckgs
-    ) {
-        if (trimmed.startsWith(ChartsWriter.VRSNS)) {
-            final String vers = trimmed.replace(ChartsWriter.VRSNS, "").trim();
-            if (pckgs.containsKey(name) && pckgs.get(name).stream().anyMatch(
-                pair -> pair.getLeft().equals(vers)
-            )) {
-                throw new IllegalStateException(
-                    String.format("Failed to write to index `%s` with version `%s`", name, vers)
-                );
-            }
-        }
-    }
-
-    /**
-     * Write remained versions of passed chart in collection in case of their existence.
-     * @param name Chart name for which remained versions are checked
-     * @param pckgs Packages collection which contains info about passed packages for
-     *  adding to index file. There is a version and chart yaml for each package.
-     * @param writer Yaml writer
-     * @throws IOException In case of exception during writing
-     */
-    private static void writeRemainedVersionsOfChart(
-        final String name,
-        final Map<String, Set<Pair<String, ChartYaml>>> pckgs,
-        final YamlWriter writer
-    ) throws IOException {
-        if (name != null && pckgs.containsKey(name)) {
-            for (final Pair<String, ChartYaml> pair : pckgs.get(name)) {
-                writer.writeLine("-", 2);
-                final String str = new IndexYamlMapping(pair.getRight().fields()).toString();
-                for (final String entry : str.split("[\\n\\r]+")) {
-                    // @checkstyle MagicNumberCheck (1 line)
-                    writer.writeLine(entry, 3);
-                }
-            }
-            pckgs.remove(name);
-        }
-    }
-
-    /**
-     * Write remained versions for all charts in collection in case of their existence.
-     * @param pckgs Packages collection which contains info about passed packages for
-     *  adding to index file. There is a version and chart yaml for each package.
-     * @param writer Yaml writer
-     */
-    private static void writeRemainedChartsAfterCopyIndex(
-        final Map<String, Set<Pair<String, ChartYaml>>> pckgs,
-        final YamlWriter writer
-    ) {
-        pckgs.forEach(
-            (chart, pairs) -> {
-                try {
-                    writer.writeLine(String.format("%s:", chart), 1);
-                    for (final Pair<String, ChartYaml> pair : pairs) {
-                        writer.writeLine("- ", 2);
-                        final String yaml;
-                        yaml = new IndexYamlMapping(pair.getRight().fields()).toString();
-                        final String[] lines = yaml.split("[\\n\\r]+");
-                        for (final String line : lines) {
-                            // @checkstyle MagicNumberCheck (1 line)
-                            writer.writeLine(line, 3);
-                        }
-                    }
-                } catch (final IOException exc) {
-                    throw  new UncheckedIOException(exc);
-                }
-            }
-        );
-        pckgs.clear();
     }
 
     /**
